Черновик NodeController

+++ node-api/src/main/java/com/github/liyasharipova/blockchain/node/api/dto/response/NonceCheckResponse.java
@@ -29,6 +29,6 @@ public class NonceCheckResponse {
      * @return blockNumber
      **/
     @JsonProperty("blockNumber")
-    private Integer blockNumber;
+    private Long blockNumber;
 }
